Moved image cropping update into Lightning Media Image.
