+++ modules/lightning_features/lightning_media/src/Update/Update218.php
@@ -2,9 +2,7 @@
 
 namespace Drupal\lightning_media\Update;
 
-use Drupal\Core\Config\ConfigFactoryInterface;
 use Drupal\Core\DependencyInjection\ContainerInjectionInterface;
-use Drupal\Core\Entity\EntityStorageInterface;
 use Drupal\Core\Extension\ModuleInstallerInterface;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
@@ -23,31 +21,13 @@ final class Update218 implements ContainerInjectionInterface {
   protected $moduleInstaller;
 
   /**
-   * The form display entity storage handler.
+   * Update218 constructor.
    *
-   * @var \Drupal\Core\Entity\EntityStorageInterface
+   * @param \Drupal\Core\Extension\ModuleInstallerInterface $module_installer
+   *   The module installer service.
    */
-  protected $formDisplayStorage;
-
-  /**
-   * The Drupal application root.
-   *
-   * @var string
-   */
-  protected $appRoot;
-
-  /**
-   * The config factory service.
-   *
-   * @var \Drupal\Core\Config\ConfigFactoryInterface
-   */
-  protected $configFactory;
-
-  public function __construct(ModuleInstallerInterface $module_installer, EntityStorageInterface $form_display_storage, $app_root, ConfigFactoryInterface $config_factory) {
+  public function __construct(ModuleInstallerInterface $module_installer) {
     $this->moduleInstaller = $module_installer;
-    $this->formDisplayStorage = $form_display_storage;
-    $this->appRoot = (string) $app_root;
-    $this->configFactory = $config_factory;
   }
 
   /**
@@ -55,54 +35,8 @@ final class Update218 implements ContainerInjectionInterface {
    */
   public static function create(ContainerInterface $container) {
     return new static(
-      $container->get('module_installer'),
-      $container->get('entity_type.manager')->getStorage('entity_form_display'),
-      $container->get('app.root'),
-      $container->get('config.factory')
+      $container->get('module_installer')
     );
-  }
-
-  /**
-   * @update
-   *
-   * @ask Do you want to enable image cropping and use it for images?
-   */
-  public function imageCropping() {
-    $this->moduleInstaller->install(['image_widget_crop']);
-
-    // Use the cropping widgets for every form display of the Image media type.
-    // This code is lifted almost directly from lightning_media_image_install().
-    $form_displays = $this->formDisplayStorage->loadByProperties([
-      'targetEntityType' => 'media',
-      'bundle' => 'image',
-    ]);
-
-    /** @var \Drupal\Core\Entity\Display\EntityFormDisplayInterface $form_display */
-    foreach ($form_displays as $form_display) {
-      $component = $form_display->getComponent('image');
-      $component['type'] = 'image_widget_crop';
-      $component['settings']['crop_list'] = ['freeform'];
-      $component['settings']['show_crop_area'] = TRUE;
-
-      if ($form_display->getMode() == 'media_browser') {
-        $component['third_party_settings']['lightning_media'] = [
-          'file_links' => FALSE,
-          'remove_button' => FALSE,
-        ];
-      }
-
-      $form_display->setComponent('image', $component);
-      $this->formDisplayStorage->save($form_display);
-    }
-
-    $cropper = 'libraries/cropper/dist/cropper.min';
-    if (file_exists($this->appRoot . "/$cropper.js")) {
-      $this->configFactory
-        ->getEditable('image_widget_crop.settings')
-        ->set('settings.library_url', "$cropper.js")
-        ->set('settings.css_url', "$cropper.css")
-        ->save();
-    }
   }
 
   /**
